services/ml-insights-service/requirements.txt
flask
scikit-learn
pandas
numpy
joblib
python-dotenv
gunicorn
requests # For calling other services
